Added two utility commands for finding duplicate and unique lines in a file
This also introduces a hacky ArgvInput which is required to allow the dispatcher
command to pass options through to the dispatched command. Without this change
you cannot specify uptions if they are not registered with the dispatcher command.
Do not rely on this class too much. I will look to removing this in the future, and
possibly replace it with a new InputDefinition that allows for this.